chore: remove outdated README for blockchain tests and refactor component imports in test-blockchain page
